+++ test/Sources.deprecated.js
@@ -1,4 +1,5 @@
 /*
+Copyright 2019 Javier Brea
 Copyright 2019 XbyOrange
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
@@ -42,7 +43,7 @@ test.describe("sources handler", () => {
     test.it("should print a trace when duplicated id is detected", () => {
       new FooOrigin("foo-id");
       new FooOrigin("foo-id");
-      test.expect(console.warn.getCall(0).args[0]).to.contain("foo-id");
+      test.expect(console.warn.getCall(2).args[0]).to.contain("foo-id");
     });
   });